fix: Chrome 85 breaks our clever download pattern
The clever download pattern we established whereby the protocol handler
would be used to start a download, and fallback to an installer if the
protocol handler was not present, is broken with Chrome 85. It turns out
that this was never explicitly supported, and was just a side-effect, so
it's unlikely to return. This makes a marginally less smooth install
experience but it should not be a significant issue, just one extra
click on first install.