fix: install link broken
The bundled installer link for Keyman keyboards for Windows
was not working if bcp47 was not specified. Furthermore, it
would default to the stable setup.exe, which is not correct
while we are in alpha or beta for 14.0, as earlier stable
versions do not support the custom filename pattern we need.