fix: bookmarklet list missing many keyboards
Fixes keymanapp/api.keyman.com#47.

This adds the version parameter to the api call in order to get all
keyboards compatible with the current version. Also tidied up global
variables for future maintainability.

There has long been a hidden feature which we have not really used, so
I will highlight out again here: you can pass in a `keyboard` and
`language` parameter into the page to get the page to show just a single
keyboard, for example:

https://keyman.com/bookmarklet/?keyboard=tibetan_direct_input&language=adi-tibt

I can see this being useful in a variety of possible situations in the
future.