feat(android): Google Play Install Referrer link
* Relates to keymanapp/keyman#5219.
* Goes with keymanapp/keyman#5230.

Add the link for Google Play Install Referrer and reorganise the install
page for Android to make the one-step install clear.

Also cleans up a few bits and pieces with the look and feel of the
install, fixing a HTML layout error and tweaking some CSS to make the
page easier to understand.