fix: allow #osk-host element to grow in keyboard details
Some keyboards use `&KMWHelpFile` to present a HTML fragment instead of
the normal KeymanWeb OSK. Those keyboards were not displaying correctly
in this view, with the content overlapping the Keyboard Details area.

Allowing the `#osk-host` element to grow mitigates this, although there
are still keyboards such as https://keyman.com/keyboards/sil_ipa that do
not have an optimal presentation.

Fixes: #573
Co-authored-by: Joshua Horton <joshua_horton@sil.org>